mindre ram
