+++ src/graphics/Sphere.h
@@ -1,11 +1,11 @@
 #pragma once
 #include "Shape.h"
+#include "Mesh.h"
 
 class Sphere : public Shape{
 public:
-    Sphere(const float& radius, const int& pointAmount);
+    Sphere(const float& radius, Mesh* mesh);
     ~Sphere();
     
     const float m_Radius;
-    const int m_PointAmount;
 };
